Task4

+++ src/by/stepanovichalena/library/service/impl/BookServiceImpl.java
@@ -5,14 +5,16 @@ import by.stepanovichalena.library.dao.exception.LibraryDAOException;
 import by.stepanovichalena.library.dao.BookDAO;
 import by.stepanovichalena.library.service.BookService;
 import by.stepanovichalena.library.service.exception.ServiceException;
+import by.stepanovichalena.library.validation.BookValidation;
+import by.stepanovichalena.library.validation.impl.BookValidator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.Collection;
 
 public class BookServiceImpl implements BookService {
-
     private static final Logger LOGGER = LogManager.getLogger(BookServiceImpl.class);
+    private BookValidation bookValidation = new BookValidator();
     private BookDAO bookDAO;
 
     public BookServiceImpl() {
@@ -24,21 +26,27 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public boolean addBook(Book book) throws ServiceException {
-        boolean result;
-        if (book == null){
-            return false;
-        }
-        try {
-            result = bookDAO.create(book);
-        } catch (LibraryDAOException e) {
-            LOGGER.error("Exception while adding the book ", e);
-            throw new ServiceException("Exception while adding the book ", e);
+        boolean result = false;
+        if (book == null) return false;
+        if (bookValidation.isBookDataValid(book.getAuthorsName(), book.getTitle())) {
+            try {
+                result = bookDAO.create(book);
+            } catch (LibraryDAOException e) {
+                LOGGER.error("Exception while adding the book ", e);
+                throw new ServiceException("Exception while adding the book ", e);
+            }
         }
         return result;
     }
 
     @Override
     public Collection<Book> find(Book book) throws ServiceException {
+        if (book == null) {
+            throw new ServiceException("Incorrect data");
+        }
+        if (!bookValidation.isBookDataValid(book.getAuthorsName(), book.getTitle())) {
+            throw new ServiceException("Incorrect data");
+        }
         try {
             return bookDAO.find(book);
         } catch (LibraryDAOException e) {
@@ -49,6 +57,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public Collection<Book> findByAuthor(String authorsName) throws ServiceException {
+        if (!bookValidation.isAuthorsNameValid(authorsName)) {
+            throw new ServiceException("Incorrect data");
+        }
         try {
             return bookDAO.findByAuthorBook(authorsName);
         } catch (LibraryDAOException e) {
@@ -59,6 +70,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public Collection<Book> findByTitle(String title) throws ServiceException {
+        if (!bookValidation.isTitleValid(title)) {
+            throw new ServiceException("Incorrect data");
+        }
         try {
             return bookDAO.findByTitle(title);
         } catch (LibraryDAOException e) {
@@ -81,26 +95,38 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public boolean delete(Book book) throws ServiceException {
-        boolean result;
-        try {
-            result = bookDAO.delete(book);
-        } catch (LibraryDAOException e) {
-            LOGGER.error("Exception in BookServiceImpl while deleting the book ", e);
-            throw new ServiceException("Exception in BookServiceImpl while deleting the book ", e);
+        boolean result = false;
+        if (book != null && bookValidation.isBookDataValid(book.getAuthorsName(), book.getTitle())) {
+            try {
+                result = bookDAO.delete(book);
+            } catch (LibraryDAOException e) {
+                LOGGER.error("Exception in BookServiceImpl while deleting the book ", e);
+                throw new ServiceException("Exception in BookServiceImpl while deleting the book ", e);
+            }
         }
         return result;
     }
 
     @Override
     public boolean deleteBookById(Integer id) throws ServiceException {
-        boolean result;
-        try {
-            result = bookDAO.deleteById(id);
-            LOGGER.info(result);
-        } catch (LibraryDAOException e) {
-            LOGGER.error("Exception in BookServiceImpl while deleting the book ", e);
-            throw new ServiceException("Exception in BookServiceImpl while deleting the book ", e);
+        boolean result = false;
+        if (id != null && bookValidation.isIdValid(id.toString())) {
+            try {
+                result = bookDAO.deleteById(id);
+                LOGGER.info(result);
+            } catch (LibraryDAOException e) {
+                LOGGER.error("Exception in BookServiceImpl while deleting the book ", e);
+                throw new ServiceException("Exception in BookServiceImpl while deleting the book ", e);
+            }
         }
         return result;
+    }
+
+    public void setBookValidation(BookValidation bookValidation) {
+        this.bookValidation = bookValidation;
+    }
+
+    public void setBookDAO(BookDAO bookDAO) {
+        this.bookDAO = bookDAO;
     }
 }
